now tells when it's safe to water the plant

+++ main.py
@@ -11,23 +11,30 @@ thresholdDown = 250
 while True:
     moisture = MoisturePin()
     millivoltsTemp = TempPin.voltage()
+    drynessStatus = False
 
     celsius = (millivoltsTemp - 500.0) / 10.0
     print("The temperature is ", celsius, "Celsius")
 
     if moisture <= thresholdDown:
-        print("Dry, Water it!")
+        print("The plant is dry!")
+        print(moisture)
+        drynessStatus = True
+
+        # Send plant moisture status to Pybytes
+        pybytes.send_signal(2, "The Plant is Dry!")
+
+    if moisture >= thresholdUp:
+        print("The plant is wet!")
         print(moisture)
 
         # Send plant moisture status to Pybytes
-        pybytes.send_signal(2, "The Plant is Dry! Water it!")
+        pybytes.send_signal(2, "The Plant is Wet!")
 
-    if moisture >= thresholdUp:
-        print("Wet, Leave it!")
-        print(moisture)
-
-    # Send temperature data to Pybytes.
-    pybytes.send_signal(1, celsius)
-    print("sending: {}".format(celsius))
+    if drynessStatus == True and celsius < 30:
+        print("The plant is ready to be watered.")
+        pybytes.send_signal(3, "Water it NOW!")
+        pybytes.send_signal(1, celsius)
+        print("sending: {}".format(celsius))
 
     time.sleep(10)
